requirements.txt
aiosqlite==0.19.0
anyio==3.6.2
asyncpg==0.27.0
black==23.3.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
